Fix segfault in encoder on platforms that use I420 CPU frame buffer types.
On MacOS (and any other platform without native texture support) a segfault will happen if the encoder thread times out waiting for the GPU to copy the native Unity frame texture to the CPU to be converted to I420 format. This timeout is currently set to 60ms and can trigger on a lag spike of GPU work. When this happens, a null buffer will be submitted to the encoder, which will then segfault attempting to access it.